requirements_server.txt
grpcio
grpcio-tools
tensorflow
grpcio_testing